LayoutTests/crypto/subtle/sign-malformed-parameters-expected.txt

Test signing with malformed parameters

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS crypto.subtle.sign() rejected promise  with TypeError: Not enough arguments.
PASS crypto.subtle.sign(1) rejected promise  with TypeError: Not enough arguments.
PASS crypto.subtle.sign(1, 2) rejected promise  with TypeError: Not enough arguments.
PASS crypto.subtle.sign("RSA-PSS", wrongKey, text) rejected promise  with TypeError: Member RsaPssParams.saltLength is required and must be an instance of unsigned long.
PASS crypto.subtle.sign("hmac", wrongKey, text) rejected promise  with InvalidAccessError: CryptoKey doesn't support signing.
PASS successfullyParsed is true

TEST COMPLETE
